feat(content): add article async outbox and audit pipeline

+++ src/hosts/CommercialNews.Api/Api/Admin/Contracts/Content/Categories/Requests/RestoreCategoryRequest.cs
@@ -2,6 +2,6 @@ namespace CommercialNews.Api.Api.Admin.Contracts.Content.Categories.Requests
 {
     public sealed class RestoreCategoryRequest
     {
-        public int ExpectedVersion { get; init; }
+        public long ExpectedVersion { get; init; }
     }
 }
