+++ src/hosts/CommercialNews.Api/Api/Admin/Contracts/Content/Tags/Requests/RestoreTagRequest.cs
@@ -2,6 +2,6 @@ namespace CommercialNews.Api.Api.Admin.Contracts.Content.Tags.Requests
 {
     public sealed class RestoreTagRequest
     {
-        public int ExpectedVersion { get; init; }
+        public long ExpectedVersion { get; init; }
     }
 }
